pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "onekommafive"
version = "0.1.0"
description = "Python API client for the 1KOMMA5° Heartbeat API"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["1komma5", "heartbeat", "energy", "home-automation", "api-client"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home Automation",
]
dependencies = [
    "PyJWT[cryptography]>=2.8",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "responses>=0.25",
]

[project.scripts]
1k5 = "onekommafive.cli:main"

[project.urls]
Repository = "https://github.com/mrebbert/1komma5-api"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
